Support embedding of libstatusnotifier
If compiled with --enable-statusnotifier=embedded, it gets compiled
inside plugins/ddb_gui_GTK*.so. This is because libstatusnotifier is not
a library you usually find on common distros.

The library does not have additional linking requirements besides the
GTK ones, so it should not harm to have it always linked in.

Warning: statusnotifier is GPLv3 so this makes GTK UI GPLv3 instead of
GPLv2+