Fix segfault in artwork cache code
The cache_cleaner_thread() function in artwork/cache.c tries
to opendir(3) covers_path, but return value is unchecked.

Later, it calls closedir(3), and if the directory did not exist
in the first place, NULL will be passed and it will segfault.

Fix by checking if covers_dir is not NULL before calling closedir().