blargg_endian: Fails to build on non-x86 with musl
Currently blargg_endian fails to detect endiness on non-x86, as ppc64le,
when not using glibc.

In my case, I am using musl on ppc64le, and deadbeef fails to compile
because the endianess is not known.

This patch check if the macros are defined before using it, other than
checking for GLIBC.