Bounds sanitization
plugins/.../nes_apu.c:

`apu_t::regs` has type `unsigned char [0x17]`, but in
`device_reset_nesapu`, iteration 24 invokes `apu_write` writing at
offset 0x17.


plugins/gtkui/plcommon.c:
playlist.c:
plugins/sc68/libsc68/conf68.c:

Unlike snprintf, with strncat the limiter on the number of bytes
written does not include the terminating null byte. Writing
`sizeof(format)` bytes plus a terminating null byte causes a string
overflow.

The second hunk in playlist.c had a few problems.
strncat ensures the destination is always null terminated, so the extra
assignemnt in conf68.c is not needed - and will additionally overflow,
if n was the size of the buffer, as when the preceding condition was
true.
If the constant d_name exceeded the length of the fullname buffer, the
unsigned n could wrap.

The intent seems to have been to copy the "sch dirname:" component of
the fullname, so I've rewritten this section to reflect this.


The warnings for src/wildmidi_lib.c seem to be false positives.


This review is by no means exhaustive. I saw compiler warnings when
building, and took a look.