tf: Ensure recursive calls have working memory
Buffer reuse is common throughout function calls, which is generally a
good thing for the most common case of small literals or minor ($add)
functions. However with more complex functions it's possible to, for
example, end up trying to use a tiny ten byte buffer - eventually for
the number of directories to go up - for complex chains of $strchr into
$div.

This adds a check for this scenario. Immediately before a function call
that would have otherwise happened with a small buffer, insert a fat
temporary buffer that can survive trimming from many levels of
recursion before a new buffer is required.