artwork_legacy: A more robust sanitization function.
This function trims spaces from the begining and end of a string. As well convert other white space types in the ascii range to a space.

Moreover on windows it uses the functions for detecting illegal characters and names, and sanitizes those.  one last thing it does on windows it makes sure the name does not end with a period.