cocoaui: fix "Show in Finder" context menu item
fixes a uri/url confusion that would prevent the menu item
from working sometimes. `[NSURL URLWithString:` would succeeds on
some path :URIs (that start with '/') and make URLs that didn't
work with activateFileViewerSelectingURLs:

The `[urls count] > 0` is a separate but related improvement,
that makes it more clear that there's nothing to show in some
cases of failure.