configure: Use AC_SYS_LARGEFILE
With musl-1.2.4 LFS support is being disabled and will later be entirely
removed where it is expected to use off_t and lseek rather than off64_t
and lseek64.

With glibc these types are also obsolete where it is expected to use the
_FILE_OFFSET_BITS define where off_t and lseek will become 64-bit
compatible versions when needed. This is set by AC_SYS_LARGEFILE.