test: restore cpp libs even if tests fail
previously `test.sh` used `set -e` to exit if compiling the tests
OR running them failed. this meant that the cpp libs had to be restored
manually, as the final line of the script was not reached. by
refactoring compilation into its own function and checking its return
code, we can exit only when compilation fails. in this event, cpp libs
are restored.